docs(storybook): align Vue capability chapters with React
