+++ apps/storybook-vue/.storybook/preview.ts
@@ -11,7 +11,39 @@ const preview: Preview = {
     a11y: { test: 'todo' },
     options: {
       storySort: {
-        order: ['Playground', 'Basics', 'Customization', 'Streaming', 'Documents', 'Performance Lab', 'QA'],
+        // Storybook statically parses this array. Keep both renderer orders aligned.
+        order: [
+          'Introduction',
+          'Playground',
+          'Basics',
+          ['Markdown Basics', 'Math', 'CJK & International Text', 'Footnotes & Definition Lists', 'Engine Plugins'],
+          'Customization',
+          [
+            'Custom Components',
+            'Metadata',
+            'URL Sanitization',
+            'Content Preprocessors',
+            'Orphan References',
+            'Theming',
+            'Extending',
+          ],
+          'Streaming',
+          [
+            'Streaming Basics',
+            'Incremental Parsing',
+            'Smooth Streaming',
+            'Streaming Cursor',
+            'Turn Taking',
+            'Error Recovery',
+          ],
+          'Documents',
+          ['Cross-Chunk Coordination', 'Definition Lifecycle'],
+          'Integrations',
+          ['Mantine'],
+          'Performance Lab',
+          ['About', '*'],
+          'QA',
+        ],
       },
     },
   },
