+++ packages/vue/stories/features/Math.stories.ts
@@ -1,0 +1,28 @@
+import { MATH_DOC } from '@ai-markdown/storybook-kit/common/fixtures';
+import type { Meta, StoryObj } from '@storybook/vue3-vite';
+import { expect, waitFor } from 'storybook/test';
+import { h } from 'vue';
+import AIMarkdown from '../../src';
+const meta: Meta = {
+  title: 'Basics/Math',
+  tags: ['autodocs'],
+  component: AIMarkdown,
+  render: (args) => ({ setup: () => () => h(AIMarkdown, { ...args, content: args.content ?? '' }) }),
+  argTypes: { content: { control: 'text' } },
+  parameters: {
+    docs: {
+      description: {
+        component:
+          'Corpus math is rendered with KaTeX. Inline and display formulas share the engine pipeline with React. The catalog imports KaTeX CSS; applications must import it too. Edit source to inspect formula changes.',
+      },
+    },
+  },
+};
+export default meta;
+type Story = StoryObj<typeof AIMarkdown>;
+export const Math: Story = {
+  args: { content: MATH_DOC },
+  play: async ({ canvasElement }) => {
+    await waitFor(() => expect(canvasElement.querySelector('.katex')).not.toBeNull());
+  },
+};
